requirements.txt
alembic==1.11.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.18
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.15.0
